move some methods into MacroAssembler
